pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agendades-scraper"
version = "0.1.0"
description = "Web scraper para eventos culturales de España - agendades.es"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Agendades Team"}
]
keywords = ["scraper", "events", "spain", "cultura"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # HTTP & Scraping
    "httpx>=0.27.0",
    "scrapy>=2.11.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",

    # Data validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "supabase>=2.3.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Date parsing
    "python-dateutil>=2.8.0",
    "babel>=2.14.0",

    # Logging & Monitoring
    "structlog>=24.1.0",
    "rich>=13.7.0",

    # Utils
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0.0",

    # LLM (Groq for enrichment)
    "groq>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
scraper = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
